requirements.txt
asgiref==3.4.1
Django==3.2.16
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.1.1
